SECURIZED !

+++ server/routes/article.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 //get all 
 
@@ -38,11 +39,16 @@ router.get("/name/:id", async (req,res) => {
 
 //create
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
         const {name, type, pic, description} = req.body
-        const newArticle = await pool.query("INSERT INTO article (article_name, article_type, article_pic, article_description) VALUES ($1, $2, $3, $4) RETURNING *", [name, type, pic, description])
-        res.json(newArticle.rows[0])
+        if (req.polyuser && req.role === "admin") {
+            const newArticle = await pool.query("INSERT INTO article (article_name, article_type, article_pic, article_description) VALUES ($1, $2, $3, $4) RETURNING *", [name, type, pic, description])
+            res.json(newArticle.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
 
     } catch (err) {
         console.error(err.message)
@@ -51,12 +57,17 @@ router.post("/", async (req,res) => {
 
 //update
 
-router.put("/id/:id", async (req,res) => {
+router.put("/id/:id", auth, async (req,res) => {
     try {
         const {id} = req.params
         const {name, type, pic, description} = req.body
-        const updateArticle = await pool.query("UPDATE article SET article_name = $2, article_type = $3, article_pic = $4, article_description = $5 WHERE article_id = $1 RETURNING *",[id, name, type, pic, description])
-        res.json(updateArticle.rows[0])
+        if (req.polyuser && req.role === "admin") {
+            const updateArticle = await pool.query("UPDATE article SET article_name = $2, article_type = $3, article_pic = $4, article_description = $5 WHERE article_id = $1 RETURNING *",[id, name, type, pic, description])
+            res.json(updateArticle.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
     } catch (err) {
         console.error(err.message)
     }
@@ -64,10 +75,15 @@ router.put("/id/:id", async (req,res) => {
 
 //delete
 
-router.delete("/id/:id", async (req,res) => {
+router.delete("/id/:id", auth, async (req,res) => {
     try {
         const {id} = req.params
-        const deleteArticle = await pool.query("DELETE FROM article WHERE article_id = $1",[id])
+        if (req.polyuser && req.role === "admin") {
+            const deleteArticle = await pool.query("DELETE FROM article WHERE article_id = $1",[id])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
     } catch (err) {
         console.error(err.message)
     }
