+++ server/routes/comment.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 //get all 
 
@@ -38,24 +39,16 @@ router.get("/article/:id", async (req,res) => {
 
 //create
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
         const {description,polyuser, article} = req.body
-        const newComment = await pool.query("INSERT INTO comment (comment_article, comment_polyuser, comment_description) VALUES ($1, $2, $3) RETURNING *", [article, polyuser, description])
-        res.json(newComment.rows[0])
-
-    } catch (err) {
-        console.error(err.message)
-    }
-})
-
-//update
-
-router.put("/id/:id", async (req,res) => {
-    try {
-        const {id} = req.params
-        const {description, likes, dislikes, polyuser, article} = req.body
-        const updateComment = await pool.query("UPDATE comment SET comment_description = $2, comment_likes = $3, comment_dislikes = $4, comment_polyuser = $5, comment_article = $6, WHERE comment_id = $1",[id, description, likes, dislikes, polyuser, article])
+        if (req.polyuser) {
+            const newComment = await pool.query("INSERT INTO comment (comment_article, comment_polyuser, comment_description) VALUES ($1, $2, $3) RETURNING *", [article, polyuser, description])
+            res.json(newComment.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
     } catch (err) {
         console.error(err.message)
     }
@@ -63,11 +56,16 @@ router.put("/id/:id", async (req,res) => {
 
 //delete
 
-router.delete("/id/:id", async (req,res) => {
+router.delete("/id/:id", auth, async (req,res) => {
     try {
         const {id} = req.params
-        const deleteComment = await pool.query("DELETE FROM comment WHERE comment_id = $1 RETURNING *",[id])
-        res.json(deleteComment.rows[0])
+        if (req.polyuser) {
+            const deleteComment = await pool.query("DELETE FROM comment WHERE comment_id = $1 RETURNING *",[id])
+            res.json(deleteComment.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
     } catch (err) {
         console.error(err.message)
     }
