+++ server/routes/rezo.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 //get all 
 
@@ -38,11 +39,16 @@ router.get("/name/:id", async (req,res) => {
 
 //create
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
         const {name, city, pic, description, date, adh, nonadh} = req.body
-        const newrezo = await pool.query("INSERT INTO rezo (rezo_name, rezo_city, rezo_pic, rezo_description, rezo_date, rezo_adh, rezo_nonadh) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *", [name, city, pic, description, date, adh, nonadh])
-        res.json(newrezo.rows[0])
+        if (req.polyuser && req.role === "admin") {
+            const newrezo = await pool.query("INSERT INTO rezo (rezo_name, rezo_city, rezo_pic, rezo_description, rezo_date, rezo_adh, rezo_nonadh) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *", [name, city, pic, description, date, adh, nonadh])
+            res.json(newrezo.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
 
     } catch (err) {
         console.error(err.message)
@@ -51,12 +57,17 @@ router.post("/", async (req,res) => {
 
 //update
 
-router.put("/id/:id", async (req,res) => {
+router.put("/id/:id", auth, async (req,res) => {
     try {
         const {id} = req.params
         const {name, city, pic, description, date, adh, nonadh} = req.body
-        const updaterezo = await pool.query("UPDATE rezo SET rezo_name = $2, rezo_city = $3, rezo_pic = $4, rezo_description = $5, rezo_date = $6, rezo_adh = $7, rezo_nonadh = $8 WHERE rezo_id = $1 RETURNING *",[id, name, city, pic, description, date, adh, nonadh])
-        res.json(updaterezo.rows[0])
+        if (req.polyuser && req.role === "admin") {
+            const updaterezo = await pool.query("UPDATE rezo SET rezo_name = $2, rezo_city = $3, rezo_pic = $4, rezo_description = $5, rezo_date = $6, rezo_adh = $7, rezo_nonadh = $8 WHERE rezo_id = $1 RETURNING *",[id, name, city, pic, description, date, adh, nonadh])
+            res.json(updaterezo.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
     } catch (err) {
         console.error(err.message)
     }
@@ -64,11 +75,16 @@ router.put("/id/:id", async (req,res) => {
 
 //delete
 
-router.delete("/id/:id", async (req,res) => {
+router.delete("/id/:id", auth, async (req,res) => {
     try {
         const {id} = req.params
-        const deleterezo = await pool.query("DELETE FROM rezo WHERE rezo_id = $1 RETURNING *",[id])
-        res.json(deleterezo.rows[0])
+        if (req.polyuser && req.role === "admin") {
+            const deleterezo = await pool.query("DELETE FROM rezo WHERE rezo_id = $1 RETURNING *",[id])
+            res.json(deleterezo.rows[0])
+        }
+        else {
+            return res.status(403).send("Not Authorized")
+        }
     } catch (err) {
         console.error(err.message)
     }
